Revert "Code review for strangler"
This reverts commit d5063567083e0348d678a938bd749e17343bcb8e.
