+++ strangler/src/main/java/com/iluwater/strangler/App.java
@@ -21,7 +21,7 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
 /**
  *
