+++ strangler/src/main/java/com/iluwater/strangler/HalfArithmetic.java
@@ -21,7 +21,7 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,10 +34,10 @@ public class HalfArithmetic {
   private static final Logger LOGGER = LoggerFactory.getLogger(HalfArithmetic.class);
   private static final String VERSION = "1.5";
 
-  private final HalfSource newSource;
-  private final OldSource oldSource;
+  private HalfSource newSource;
+  private OldSource oldSource;
 
-  public HalfArithmetic(HalfSource newSource, OldSource oldSource) {
+  public HalfArithmetic(final HalfSource newSource, final OldSource oldSource) {
     this.newSource = newSource;
     this.oldSource = oldSource;
   }
@@ -47,7 +47,7 @@ public class HalfArithmetic {
    * @param nums numbers need to add together
    * @return accumulate sum
    */
-  public int sum(int... nums) {
+  public int sum(final int... nums) {
     LOGGER.info("Arithmetic sum {}", VERSION);
     return newSource.accumulateSum(nums);
   }
@@ -57,7 +57,7 @@ public class HalfArithmetic {
    * @param nums numbers need to multiply together
    * @return accumulate multiplication
    */
-  public int mul(int... nums) {
+  public int mul(final int... nums) {
     LOGGER.info("Arithmetic mul {}", VERSION);
     return oldSource.accumulateMul(nums);
   }
@@ -67,7 +67,7 @@ public class HalfArithmetic {
    * @param nums numbers need to check
    * @return  if has any zero, return true, else, return false
    */
-  public boolean ifHasZero(int... nums) {
+  public boolean ifHasZero(final int... nums) {
     LOGGER.info("Arithmetic check zero {}", VERSION);
     return !newSource.ifNonZero(nums);
   }
