+++ strangler/src/main/java/com/iluwater/strangler/HalfSource.java
@@ -21,31 +21,34 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
+import java.util.Arrays;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * Test methods in NewSource
+ * Source under development. Replace part of old source and has added some new features.
  */
-public class NewSourceTest {
-  private static final NewSource source = new NewSource();
+public class HalfSource {
+  private static final Logger LOGGER = LoggerFactory.getLogger(HalfSource.class);
+  private static final  String VERSION = "1.5";
 
-  @Test
-  public void testAccumulateSum() {
-    assertEquals(0, source.accumulateSum(-1, 0, 1));
+  /**
+   * Implement accumulate sum with new technique.
+   * Replace old one in {@link OldSource}
+   */
+  public int accumulateSum(final int... nums) {
+    LOGGER.info("Source module {}", VERSION);
+    return Arrays.stream(nums).reduce(0, Integer::sum);
   }
 
-  @Test
-  public void testAccumulateMul() {
-    assertEquals(0, source.accumulateMul(-1, 0, 1));
-  }
-
-  @Test
-  public void testIfNonZero() {
-    assertFalse(source.ifNonZero(-1, 0, 1));
+  /**
+   * Check if all number is not zero.
+   * New feature.
+   */
+  public boolean ifNonZero(final int... nums) {
+    LOGGER.info("Source module {}", VERSION);
+    return Arrays.stream(nums).allMatch(num -> num != 0);
   }
 }
