+++ strangler/src/main/java/com/iluwater/strangler/OldSource.java
@@ -21,25 +21,39 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * Test methods in OldArithmetic
+ * Old source with techniques out of date.
  */
-class OldArithmeticTest {
-  private static final OldArithmetic arithmetic = new OldArithmetic(new OldSource());
+public class OldSource {
+  private static final Logger LOGGER = LoggerFactory.getLogger(OldSource.class);
+  private static final  String VERSION = "1.0";
 
-  @Test
-  public void testSum() {
-    assertEquals(0, arithmetic.sum(-1, 0, 1));
+  /**
+   * Implement accumulate sum with old technique.
+   */
+  public int accumulateSum(final int... nums) {
+    LOGGER.info("Source module {}", VERSION);
+    var sum = 0;
+    for (final var num : nums) {
+      sum += num;
+    }
+    return sum;
   }
 
-  @Test
-  public void testMul() {
-    assertEquals(0, arithmetic.mul(-1, 0, 1));
+  /**
+   * Implement accumulate multiply with old technique.
+   */
+  public int accumulateMul(final int... nums) {
+    LOGGER.info("Source module {}", VERSION);
+    var sum = 1;
+    for (final var num : nums) {
+      sum *= num;
+    }
+    return sum;
   }
 }
