+++ strangler/src/test/java/com/iluwater/strangler/HalfArithmeticTest.java
@@ -21,7 +21,7 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
 import org.junit.jupiter.api.Test;
 
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.*;
  * Test methods in HalfArithmetic
  */
 class HalfArithmeticTest {
-  private static final HalfArithmetic arithmetic = new HalfArithmetic(new HalfSource(), new OldSource());
+  private static HalfArithmetic arithmetic = new HalfArithmetic(new HalfSource(), new OldSource());
 
   @Test
   public void testSum() {
