+++ strangler/src/test/java/com/iluwater/strangler/HalfSourceTest.java
@@ -21,16 +21,26 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
 import org.junit.jupiter.api.Test;
 
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 /**
- * Application test
+ * Test methods in HalfSource
  */
-public class AppTest {
+public class HalfSourceTest {
+  private static HalfSource source = new HalfSource();
+
   @Test
-  public void test() {
-    App.main(new String[]{});
+  public void testAccumulateSum() {
+    assertEquals(0, source.accumulateSum(-1, 0, 1));
+  }
+
+  @Test
+  public void testIfNonZero() {
+    assertFalse(source.ifNonZero(-1, 0, 1));
   }
 }
