+++ strangler/src/test/java/com/iluwater/strangler/OldSourceTest.java
@@ -21,34 +21,25 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.strangler;
+package com.iluwater.strangler;
 
-import java.util.Arrays;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * Source under development. Replace part of old source and has added some new features.
+ * Test methods in OldSource
  */
-public class HalfSource {
-  private static final Logger LOGGER = LoggerFactory.getLogger(HalfSource.class);
-  private static final  String VERSION = "1.5";
+public class OldSourceTest {
+  private static OldSource source = new OldSource();
 
-  /**
-   * Implement accumulate sum with new technique.
-   * Replace old one in {@link OldSource}
-   */
-  public int accumulateSum(int... nums) {
-    LOGGER.info("Source module {}", VERSION);
-    return Arrays.stream(nums).reduce(0, Integer::sum);
+  @Test
+  public void testAccumulateSum() {
+    assertEquals(0, source.accumulateSum(-1, 0, 1));
   }
 
-  /**
-   * Check if all number is not zero.
-   * New feature.
-   */
-  public boolean ifNonZero(int... nums) {
-    LOGGER.info("Source module {}", VERSION);
-    return Arrays.stream(nums).allMatch(num -> num != 0);
+  @Test
+  public void testAccumulateMul() {
+    assertEquals(0, source.accumulateMul(-1, 0, 1));
   }
 }
